[Refactor] #125 - Factory 패턴 간단하게 적용, TestVC -> RegisterNormalViewController로 파일 변경
