+++ WSSiOS/WSSiOS/Resource/Extensions/UIViewController+.swift
@@ -66,9 +66,16 @@ extension UIViewController {
         self.navigationItem.rightBarButtonItem = right != nil ? UIBarButtonItem(customView: right!) : nil
     }
     
+    func pushToRegisterNormalViewController(novelId: Int) {
+        let registerNormalViewController = ModuleFactory.shared.makeRegisterNormalViewController(novelId: novelId)
+        self.navigationController?.pushViewController(registerNormalViewController,
+                                                      animated: true)
+    }
+    
     func pushToRegisterSuccessViewController(userNovelId: Int) {
-        self.navigationController?.pushViewController(RegisterSuccessViewController(userNovelId: userNovelId),
-                                                       animated: true)
+        let successViewController = ModuleFactory.shared.makeRegisterSuccessViewController(userNovelId: userNovelId)
+        self.navigationController?.pushViewController(successViewController,
+                                                      animated: true)
     }
     
     func moveToNovelDetailViewController(userNovelId: Int) {
@@ -93,17 +100,6 @@ extension UIViewController {
                     memoService: DefaultMemoService()),
                 memoId: memoId
             ), animated: true)
-    }
-    
-    func pushToRegisterNormalViewController(novelId: Int) {
-        self.navigationController?.pushViewController(
-            RegisterNormalViewController(
-                novelRepository: DefaultNovelRepository(
-                    novelService: DefaultNovelService()),
-                userNovelRepository: DefaultUserNovelRepository(
-                    userNovelService: DefaultUserNovelService()),
-                novelId: novelId),
-            animated: true)
     }
     
     func pushToMemoEditViewController(userNovelId: Int? = nil, memoId: Int? = nil, novelTitle: String, novelAuthor: String, novelImage: String, memoContent: String? = nil) {
